correção timezone

+++ checkout-backend/src/main/java/br/com/foxi/checkout/security/SecurityConfig.java
@@ -1,5 +1,7 @@
 package br.com.foxi.checkout.security;
 
+import java.util.Date;
+
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 
@@ -8,7 +10,7 @@ import org.springframework.context.annotation.Configuration;
 public class SecurityConfig {
     public static String PREFIX;
     public static String KEY;
-    public static Long EXPIRATION;
+    public static int EXPIRATION;
 
     public void setPrefix(String prefix){
         PREFIX = prefix;
@@ -16,7 +18,8 @@ public class SecurityConfig {
     public void setKey(String key){
         KEY = key;
     }
-    public void setExpiration(Long expiration){
+
+	public void setExpiration(int expiration){
         EXPIRATION = expiration;
     }
 }
